7.x builds should now work with VS2010
The 7.x projects were referencing the VS2012 (11.0) version of
Microsoft.VisualStudio.ComponentHost.dll, and this stopped them loading
in VS2010. They now reference the VS2010 (10.0) version of the dll,
which is included in the commit so we can still build on a system with
only VS2012. VS2012 redirects the assembly and loads the 11.0 version