No longer references dlls in bin folders
ReSharper tracks the file location of each usage of each assembly. The
list of file locations is in an undefined order, and the reference
quick-fix uses the first. If the first is in a bin folder, we don't find
it in NuGet. We now check all file locations for a NuGet package. Fixes
#2

Also tested in ReSharper 7.1, and bumped version to 1.0. Fixes #5