Line endings
Don't understand this. Edited the file on GitHub, now local file is saying that it's changed, but is CRLF locally. When GitHub saved the file, it should have been LF in the object DB. Getting a local copy would have converted it to CRLF, and the local copy would commit as LF, so there should be no mismatch. Has GitHub saved it into the object DB as CRLF?